@@ -1,87 +1,47 @@
 {
-  "name": "BettrDash",
+  "name": "desktop",
   "version": "0.1.0",
   "private": true,
-  "homepage": "./",
-  "main": "./build/electron/main.js",
-  "author": "Elias Wambugu",
+  "main": "./build/main.js",
+  "dist": "electron-builder",
+  "build": {
+    "appId": "com.bettrdash.desktop",
+    "productName": "BettrDash",
+    "directories": {
+      "output": "build"
+    }
+  },
+  "electron": "electron .",
   "dependencies": {
-    "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "2.3.6",
-    "@emotion/react": "^11.0.0",
-    "@emotion/styled": "^11.0.0",
-    "@testing-library/jest-dom": "^5.9.0",
-    "@testing-library/react": "^10.2.1",
-    "@testing-library/user-event": "^12.0.2",
-    "@types/jest": "^25.0.0",
-    "@types/node": "^18.7.23",
-    "@types/react": "18",
-    "@types/react-dom": "^16.9.0",
-    "@uploadcare/react-widget": "^2.0.0",
-    "add": "^2.0.6",
-    "axios": "^0.26.1",
-    "electron-reload": "^2.0.0-alpha.1",
-    "env-cmd": "^10.1.0",
-    "framer-motion": "^4.0.0",
-    "history": "^5.3.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-icons": "^4.3.1",
-    "react-query": "^3.35.0",
-    "react-router-dom": "^6.3.0",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "typescript": "^4.3.5",
-    "web-vitals": "^0.2.2"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build:desktop": "npm run electron:build && tsc -p electron && electron-builder -c.extraMetadata.main=build/electron/main.js",
-    "start:electron": "concurrently \"cross-env BROWSER=none npm start\" \"wait-on http://localhost:3000 && tsc -p electron -w\" \"wait-on http://localhost:3000 && tsc -p electron && electron .\"",
-    "electron:build": "react-scripts build",
-    "desktop": "npm run start:electron",
+    "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "build-qa": "env-cmd -f .env react-scripts build"
+    "electron": "electron .",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": ["react-app", "react-app/jest"]
   },
   "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
+    "production": [">0.2%", "not dead", "not op_mini all"],
     "development": [
       "last 1 chrome version",
       "last 1 firefox version",
       "last 1 safari version"
     ]
   },
-  "build": {
-    "extends": null,
-    "appId": "com.eliaswambugu.bettrdash",
-    "files": [
-      "build/**/*",
-      "node_modules/**/*",
-      "package.json",
-      "dist/**/*"
-    ],
-    "directories": {
-      "buildResources": "assets"
-    }
-  },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-syntax-flow": "^7.14.5",
-    "@babel/plugin-transform-react-jsx": "^7.14.9",
-    "@chakra-ui/system": ">=1.0.0",
-    "@testing-library/dom": ">=7.21.4",
-    "@types/node": "^18.7.23",
-    "concurrently": "^7.4.0",
-    "copyfiles": "^2.4.1",
-    "cross-env": "^7.0.3",
-    "electron": "20.0.0",
-    "electron-builder": "23.3.3",
-    "electron-devtools-installer": "^3.2.0",
-    "electron-is-dev": "^2.0.0",
-    "wait-on": "^6.0.1"
+    "concurrently": "^8.2.2",
+    "electron": "^28.2.2",
+    "electron-is-dev": "^3.0.1"
   }
 }
