@@ -1,7 +1,7 @@
 import { app, BrowserWindow } from "electron";
 import * as path from "path";
-import * as isDev from 'electron-is-dev'
-const url = require('url');
+import * as isDev from "electron-is-dev";
+// import {url} from 'url';
 
 const createWindow = () => {
   const win = new BrowserWindow({
@@ -17,7 +17,7 @@ const createWindow = () => {
   });
   if (app.isPackaged) {
     // win.loadURL(`file://${path.join(__dirname, "../index.html")}`);
-    win.loadURL('file://' + path.join(__dirname, "../index.html"));
+    win.loadURL("file://" + path.join(__dirname, "../index.html"));
   } else {
     // win.loadURL(url.format({
     //   pathname: path.join(__dirname, "../index.html"),
@@ -33,7 +33,7 @@ const createWindow = () => {
         "..",
         "node_modules",
         ".bin",
-        "electron" + (process.platform === "win32" ? ".cmd" : "")
+        "electron" + (process.platform === "win32" ? ".cmd" : ""),
       ),
       forceHardReset: true,
       hardResetMethod: "exit",
